add iterator impl
